% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{create_columns}
\alias{create_columns}
\title{Helper function to create columns configuration.}
\usage{
create_columns(
  df,
  colNames,
  columnHeaders,
  readOnly,
  hide,
  fixedCols,
  columnOrder,
  dropDown
)
}
\arguments{
\item{df}{Data frame containing the data.}

\item{colNames}{Column names.}

\item{columnHeaders}{Custom column headers.}

\item{readOnly}{Read-only columns.}

\item{hide}{Columns to hide.}

\item{fixedCols}{Columns to fix.}

\item{columnOrder}{Desired column order.}

\item{dropDown}{List of dropdown options.}
}
\description{
Helper function to create columns configuration.
}
\keyword{internal}
