% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{get_col_indices}
\alias{get_col_indices}
\title{Helper function to validate and get column indices.}
\usage{
get_col_indices(cols, all_cols)
}
\arguments{
\item{cols}{Columns specified as indices or names.}

\item{all_cols}{All available column names.}
}
\description{
Helper function to validate and get column indices.
}
\keyword{internal}
