% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{render_tabulator}
\alias{render_tabulator}
\title{Render a Tabulator widget}
\usage{
render_tabulator(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Tabulator configuration}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\description{
Render a Tabulator widget
}
