% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{Create a Tabulator widget}
\usage{
tabulator(
  data,
  options = list(),
  elementId = NULL,
  readOnly = FALSE,
  columnHeaders = NULL,
  hide = NULL,
  fixedCols = NULL,
  stretched = c("all", "last", "none"),
  dropDown = list(),
  css = NULL,
  add_selector_bar = FALSE,
  add_select_column = FALSE,
  return_select_column = FALSE,
  return_select_column_name = "row_select",
  columnOrder = NULL,
  columns = NULL
)
}
\arguments{
\item{data}{A data frame or matrix, or NULL for empty table}

\item{options}{A list of Tabulator.js options}

\item{elementId}{The ID of the element}

\item{readOnly}{A boolean for the whole table, or a vector of column ids or names (no mixing)}

\item{columnHeaders}{Custom column headers}

\item{hide}{A vector of column ids or names to hide}

\item{fixedCols}{A vector of column ids or names, columns will be fixed from first to the specified ones}

\item{stretched}{Stretching mode for columns (default: "all")}

\item{dropDown}{A list of dropdown options for specific columns}

\item{css}{Optional path to custom CSS file}

\item{add_selector_bar}{Boolean to add a selector bar (default: FALSE)}

\item{add_select_column}{Boolean to add a select column (default: FALSE)}

\item{return_select_column}{Boolean to include selection status in returned data (default: FALSE)}

\item{return_select_column_name}{Name for the returned selection column (default: "row_select")}

\item{columnOrder}{A character vector specifying the desired column order (default: NULL)}

\item{columns}{Manual columns setting. If set, auto creation is skipped.}
}
\description{
Create a Tabulator widget
}
