% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_add_rows}
\alias{tabulator_add_rows}
\title{Add rows to a Tabulator table}
\usage{
tabulator_add_rows(proxy, data, position = "bottom")
}
\arguments{
\item{proxy}{A Tabulator proxy object}

\item{data}{A list or data frame containing the rows to add}

\item{position}{Optional position to insert rows ("top" or "bottom", default: "bottom")}
}
\description{
Add rows to a Tabulator table
}
