% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_output}
\alias{tabulator_output}
\title{Create a Tabulator output element}
\usage{
tabulator_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The ID of the output element}

\item{width}{The width of the element (default: '100%')}

\item{height}{The height of the element (default: '400px')}
}
\description{
Create a Tabulator output element
}
