% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_proxy}
\alias{tabulator_proxy}
\title{Create a Tabulator proxy object}
\usage{
tabulator_proxy(input_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{input_id}{The ID of the Tabulator input element}

\item{session}{The Shiny session object}
}
\description{
Create a Tabulator proxy object
}
