% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_remove_first_row}
\alias{tabulator_remove_first_row}
\title{Remove the first row from a Tabulator table}
\usage{
tabulator_remove_first_row(proxy)
}
\arguments{
\item{proxy}{A Tabulator proxy object}
}
\description{
Remove the first row from a Tabulator table
}
