% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_remove_rows}
\alias{tabulator_remove_rows}
\title{Remove specific rows from a Tabulator table}
\usage{
tabulator_remove_rows(proxy, row_ids)
}
\arguments{
\item{proxy}{A Tabulator proxy object}

\item{row_ids}{A vector of row IDs to remove}
}
\description{
Remove specific rows from a Tabulator table
}
