% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_to_df}
\alias{tabulator_to_df}
\title{Convert raw message to data.frame}
\usage{
tabulator_to_df(message)
}
\arguments{
\item{message}{Message from client}
}
\description{
Convert raw message to data.frame
}
