% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_trigger_data}
\alias{tabulator_trigger_data}
\title{Trigger data input ( refresh )}
\usage{
tabulator_trigger_data(proxy)
}
\arguments{
\item{proxy}{A Tabulator proxy object}
}
\description{
Trigger data input ( refresh )
}
