% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_update_data}
\alias{tabulator_update_data}
\title{Update Tabulator data}
\usage{
tabulator_update_data(proxy, data, chunk_size = 1000)
}
\arguments{
\item{proxy}{A Tabulator proxy object}

\item{data}{A data frame with updated data}

\item{chunk_size}{The number of rows to update in each chunk (default: 1000)}
}
\description{
Update Tabulator data
}
