% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator_update_where}
\alias{tabulator_update_where}
\title{Update Tabulator values using a conditional system}
\usage{
tabulator_update_where(
  proxy,
  col,
  value,
  whereCol,
  whereValue,
  operator = "==",
  chunk_size = 1000
)
}
\arguments{
\item{proxy}{A Tabulator proxy object}

\item{col}{The column to update}

\item{value}{The new value to set}

\item{whereCol}{The column to check for the condition}

\item{whereValue}{The value to compare against in the condition}

\item{operator}{The comparison operator (e.g., ">", "<", "==", etc.)}

\item{chunk_size}{The number of rows to update in each chunk (default: 1000)}
}
\description{
Update Tabulator values using a conditional system
}
